+++ biojava-structure/src/main/java/org/biojava/nbio/structure/Atom.java
@@ -240,7 +240,8 @@ public interface Atom extends Cloneable, PDBRecord {
 	public void setBonds(List<Bond> bonds);
 
 
-	/** Test if another atom has a bond to this atom
+	/**
+	 * Test if another atom has a bond to this atom
 	 *
 	 * @param other
 	 * @return